GUI z-level refactor. The GUI widgets are now stacked on z levels.
Note: Some things are still broken, most notably the text fields on any
popup GUIs currently render behind the popup GUI. This will be fixed
when the text fields are added as malilib widgets in the coming commits.

+++ src/main/java/fi/dy/masa/malilib/gui/GuiTextFieldGeneric.java
@@ -53,7 +53,7 @@ public class GuiTextFieldGeneric extends GuiTextField
         }
     }
 
-    public GuiTextFieldGeneric setZLevel(float zLevel)
+    public GuiTextFieldGeneric setZLevel(int zLevel)
     {
         this.zLevel = zLevel;
         return this;
